refactor: codacy driven corrections; removed else statements, early exits, introduced helper traits
Refs #17
